requirements.txt
# AI Agent — Python Dependencies
# Install with: pip install -r requirements.txt

# Core AI
anthropic>=0.34.0
groq
pyautogui
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP requests (for web search, weather tools)
requests>=2.32.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.9.0

# Optional: for testing
pytest>=8.0.0
httpx>=0.27.0   # async test client for FastAPI

